Move most logic of movePiece() to KifuTree model
